refactor: 테스트코드 리팩토링
