+++ app/src/androidTest/java/woowacourse/movie/MovieScreenTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.java
+package woowacourse.movie
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
@@ -10,11 +10,10 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import woowacourse.movie.R
 import woowacourse.movie.presentation.screen.MovieScreenActivity
 
 @RunWith(AndroidJUnit4::class)
-class MovieListTest {
+class MovieScreenTest {
     @get:Rule
     val activityRule = ActivityScenarioRule(MovieScreenActivity::class.java)
