+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:name=".presentation.reservation.MovieReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.detail.MovieDetailActivity"
+            android:name=".presentation.detail.TicketDetailActivity"
             android:exported="false" />
     </application>
