+++ app/src/main/res/layout/activity_movie_reservation_complete.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center_horizontal"
-    tools:context=".presentation.detail.MovieDetailActivity"
+    tools:context=".presentation.detail.TicketDetailActivity"
     android:orientation="vertical">
     <LinearLayout
         android:background="#9160F9"
